debian/control: add python-is-python2 as build dependency
The provided virtual package names are generated by the
debian/virtual-packages.py script, which uses a shebang with
'/usr/bin/python'. By default there is no python binary on ubuntu focal,
that is why we need the python-is-python2 package.
Whithout this modification the build process did not fail and the
provided virtual package names were not generated, which caused an
installation problem with the packages that depend on some virtual
package from ply.